style: ProgressBar 개별 컴포넌트로 분리
