+++ app/components/flip-card.tsx
@@ -16,17 +16,8 @@ export default function FlipCard({ duration = 100, children, ...props }: Props) 
   return (
     <div css={[cardContainerCss, flip && flipCss]} {...props}>
       <div css={[innerCss, flip && flipCss]}>
-        <div
-          css={[
-            cardCss,
-            css`
-              background-color: white;
-            `,
-          ]}
-        >
-          {children}
-        </div>
-        <div css={[backfaceCss, cardCss]}>뒷면</div>
+        <div css={[cardCss]}>{children}</div>
+        <div css={[cardCss, backfaceCss]}>뒷면</div>
       </div>
     </div>
   );
@@ -56,8 +47,8 @@ const cardCss = css`
   height: 100%;
 
   background-color: white;
-  -webkit-backface-visibility: hidden;
   backface-visibility: hidden;
+  -webkit-backface-visibility: hidden;
 `;
 
 const flipCss = css`
